Support template file with external contents
This adds support for loading a template's files on-demand instead of embedding them all into the templates file itself. The code that creates the templates file is also changed to copy files over to `dist/templates/[template name]`, as is the code for loading files to properly resolve paths relative to the templates file.