Fix process chrdir
Was setting this.cwd which works in some situations but not in others
where this isn't getting bound to process. So instead we now
set cwd directly on process but this causes recursion since resolve
uses cwd. Resolving the path ahead of time and returning the cwd
fixes this recursion.